alerts-viewer-app/requirements.txt
Flask==3.0.0
kafka-python==2.0.2
